# api/docs/swagger.yaml
basePath: /v1
definitions:
...
  responses.BadRequest:
    properties:
      data:
        example: The request body is not a valid JSON string
        type: string
      message:
        example: The request isn't properly formed
        type: string
      status:
        example: error
        type: string
    required:
      - data
      - message
      - status
    type: object
...
  responses.InternalServerError:
    properties:
      message:
        example: We ran into an internal error while handling the request.
        type: string
      status:
        example: error
        type: string
    required:
      - message
      - status
    type: object
...
  responses.MessagesResponse:
    properties:
      data:
        items:
          $ref: "#/definitions/entities.Message"
        type: array
      message:
        example: item created successfully
        type: string
      status:
        example: success
        type: string
    required:
      - data
      - message
      - status
    type: object
...
  responses.Unauthorized:
    properties:
      data:
        example: Make sure your API key is set in the [X-API-Key] header in the request
        type: string
      message:
        example: You are not authorized to carry out this request.
        type: string
      status:
        example: error
        type: string
    required:
      - data
      - message
      - status
    type: object
  responses.UnprocessableEntity:
    properties:
      data:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        example: validation errors while sending message
        type: string
      status:
        example: error
        type: string
    required:
      - data
      - message
      - status
    type: object
...
host: api.httpsms.com
info:
  contact:
    email: support@httpsms.com
    name: HTTP SMS
  description:
    API to send SMS messages using android [SmsManager](https://developer.android.com/reference/android/telephony/SmsManager)
    via HTTP
  license:
    name: AGPL-3.0
    url: https://raw.githubusercontent.com/NdoleStudio/http-sms-manager/main/LICENSE
  title: HTTP SMS API
  version: "1.0"
paths:
...
  /messages:
    get:
      consumes:
        - application/json
      description:
        Get list of messages which are sent between 2 phone numbers. It
        will be sorted by timestamp in descending order.
      parameters:
        - default: "+18005550199"
          description: the owner's phone number
          in: query
          name: owner
          required: true
          type: string
        - default: "+18005550100"
          description: the contact's phone number
          in: query
          name: contact
          required: true
          type: string
        - description: number of messages to skip
          in: query
          minimum: 0
          name: skip
          type: integer
        - description: filter messages containing query
          in: query
          name: query
          type: string
        - description: number of messages to return
          in: query
          maximum: 20
          minimum: 1
          name: limit
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/responses.MessagesResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/responses.BadRequest"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/responses.Unauthorized"
        "422":
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/responses.UnprocessableEntity"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/responses.InternalServerError"
      security:
        - ApiKeyAuth: []
      summary: Get messages which are sent between 2 phone numbers
      tags:
        - Messages
  /messages/{messageID}:
    delete:
      consumes:
        - application/json
      description:
        Delete a message from the database and removes the message content
        from the list of threads.
      parameters:
        - default: 32343a19-da5e-4b1b-a767-3298a73703ca
          description: ID of the message
          in: path
          name: messageID
          required: true
          type: string
      produces:
        - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: "#/definitions/responses.NoContent"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/responses.BadRequest"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/responses.Unauthorized"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/responses.NotFound"
        "422":
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/responses.UnprocessableEntity"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/responses.InternalServerError"
      security:
        - ApiKeyAuth: []
      summary: Delete a message from the database.
      tags:
        - Messages
  /messages/{messageID}/events:
    post:
      consumes:
        - application/json
      description:
        Use this endpoint to send events for a message when it is failed,
        sent or delivered by the mobile phone.
      parameters:
        - default: 32343a19-da5e-4b1b-a767-3298a73703ca
          description: ID of the message
          in: path
          name: messageID
          required: true
          type: string
        - description: Payload of the event emitted.
          in: body
          name: payload
          required: true
          schema:
            $ref: "#/definitions/requests.MessageEvent"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/responses.MessageResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/responses.BadRequest"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/responses.Unauthorized"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/responses.NotFound"
        "422":
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/responses.UnprocessableEntity"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/responses.InternalServerError"
      security:
        - ApiKeyAuth: []
      summary: Upsert an event for a message on the mobile phone
      tags:
        - Messages
  /messages/bulk-send:
    post:
      consumes:
        - application/json
      description: Add bulk SMS messages to be sent by the android phone
      parameters:
        - description: Bulk send message request payload
          in: body
          name: payload
          required: true
          schema:
            $ref: "#/definitions/requests.MessageBulkSend"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/responses.MessagesResponse"
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/responses.BadRequest"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/responses.Unauthorized"
        "422":
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/responses.UnprocessableEntity"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/responses.InternalServerError"
      security:
        - ApiKeyAuth: []
      summary: Send bulk SMS messages
      tags:
        - Messages
  /messages/calls/missed:
    post:
      consumes:
        - application/json
      description:
        This endpoint is called by the httpSMS android app to register
        a missed call event on the mobile phone.
      parameters:
        - description: Payload of the missed call event.
          in: body
          name: payload
          required: true
          schema:
            $ref: "#/definitions/requests.MessageCallMissed"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/responses.MessageResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/responses.BadRequest"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/responses.Unauthorized"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/responses.NotFound"
        "422":
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/responses.UnprocessableEntity"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/responses.InternalServerError"
      security:
        - ApiKeyAuth: []
      summary: Register a missed call event on the mobile phone
      tags:
        - Messages
  /messages/outstanding:
    get:
      consumes:
        - application/json
      description: Get an outstanding message to be sent by an android phone
      parameters:
        - default: 32343a19-da5e-4b1b-a767-3298a73703cb
          description: The ID of the message
          in: query
          name: message_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/responses.MessageResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/responses.BadRequest"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/responses.Unauthorized"
        "422":
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/responses.UnprocessableEntity"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/responses.InternalServerError"
      security:
        - ApiKeyAuth: []
      summary: Get an outstanding message
      tags:
        - Messages
  /messages/receive:
    post:
      consumes:
        - application/json
      description: Add a new message received from a mobile phone
      parameters:
        - description: Received message request payload
          in: body
          name: payload
          required: true
          schema:
            $ref: "#/definitions/requests.MessageReceive"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/responses.MessageResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/responses.BadRequest"
        "422":
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/responses.UnprocessableEntity"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/responses.InternalServerError"
      security:
        - ApiKeyAuth: []
      summary: Receive a new SMS message from a mobile phone
      tags:
        - Messages
  /messages/search:
    get:
      consumes:
        - application/json
      description:
        This returns the list of all messages based on the filter criteria
        including missed calls
      parameters:
        - default: +18005550199,+18005550100
          description: the owner's phone numbers
          in: query
          name: owners
          required: true
          type: string
        - description: number of messages to skip
          in: query
          minimum: 0
          name: skip
          type: integer
        - description: filter messages containing query
          in: query
          name: query
          type: string
        - description: number of messages to return
          in: query
          maximum: 200
          minimum: 1
          name: limit
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/responses.MessagesResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/responses.BadRequest"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/responses.Unauthorized"
        "422":
          description: Unprocessable Entity
          schema:
            $ref: "#/definitions/responses.UnprocessableEntity"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/responses.InternalServerError"
      security:
        - ApiKeyAuth: []
      summary: Search all messages of a user
      tags:
        - Messages
...
schemes:
  - https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-api-Key
    type: apiKey
swagger: "2.0"